+++ config/config.exs
@@ -3,7 +3,7 @@
 #
 # This configuration file is loaded before any dependency and
 # is restricted to this project.
-use Mix.Config
+import Config
 
 # Environment
 config :td_bg, :env, Mix.env()
@@ -43,7 +43,6 @@ config :phoenix, :json_library, Jason
 config :phoenix_swagger, json_library: Jason
 
 config :td_bg, TdBg.Auth.Guardian,
-  # optional
   allowed_algos: ["HS512"],
   issuer: "tdauth",
   ttl: {1, :hours},