+++ lib/td_bg/auth/error_handler.ex
@@ -2,8 +2,17 @@ defmodule TdBg.Auth.ErrorHandler do
   @moduledoc false
   import Plug.Conn
 
-  def auth_error(conn, {type, _reason}, _opts) do
+  def unauthorized(conn) do
+    conn
+    |> auth_error({:unauthorized, nil})
+    |> halt()
+  end
+
+  def auth_error(conn, {type, _reason}, _opts \\ []) do
     body = Jason.encode!(%{message: to_string(type)})
-    send_resp(conn, 401, body)
+
+    conn
+    |> put_resp_content_type("application/json")
+    |> send_resp(:unauthorized, body)
   end
 end