+++ priv/repo/migrations/20220720105543_modify_domains_description_to_text.exs
@@ -1,0 +1,9 @@
+defmodule TdBg.Repo.Migrations.ModifyDomainsDescriptionToText do
+  use Ecto.Migration
+
+  def change do
+    alter table(:domains) do
+      modify(:description, :text, from: :string)
+    end
+  end
+end